+++ app/src/main/java/io/legado/app/help/BookHelp.kt
@@ -1,5 +1,6 @@
 package io.legado.app.help
 
+import android.net.Uri
 import io.legado.app.constant.AppPattern
 import io.legado.app.constant.EventBus
 import io.legado.app.data.appDb
@@ -21,9 +22,9 @@ import kotlin.math.max
 import kotlin.math.min
 
 object BookHelp {
-    const val cacheFolderName = "book_cache"
+    private const val cacheFolderName = "book_cache"
     private const val cacheImageFolderName = "images"
-    val downloadDir: File = appCtx.externalFilesDir
+    private val downloadDir: File = appCtx.externalFilesDir
     private val downloadImages = CopyOnWriteArraySet<String>()
 
     fun clearCache() {
@@ -53,6 +54,28 @@ object BookHelp {
                 }
             }
         }
+    }
+
+    fun getEpubFile(book: Book,): File {
+        val file = FileUtils.getFile(
+            downloadDir,
+            cacheFolderName,
+            book.getFolderName(),
+            "index.epubx"
+        )
+        if(!file.exists()){
+            val input = if (book.bookUrl.isContentScheme()) {
+                val uri = Uri.parse(book.bookUrl)
+                appCtx.contentResolver.openInputStream(uri)
+            } else {
+                File(book.bookUrl).inputStream()
+            }
+            if (input != null) {
+                FileUtils.writeInputStream(file, input)
+            }
+
+        }
+        return file
     }
 
     suspend fun saveContent(book: Book, bookChapter: BookChapter, content: String) {
